Afegit WTForms

+++ my_app/routes_main.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, redirect, url_for, render_template, request
 from .models import Item, Store
+from .forms import ItemForm, DeleteForm
 from . import db_manager as db
 
 # Blueprint
@@ -21,20 +22,17 @@ def items_list():
 def items_update(item_id):
     # select amb 1 resultat
     item = db.session.query(Item).filter(Item.id == item_id).one()
-    
-    if request.method == 'GET':
-        # select que retorna una llista de resultats
-        stores = db.session.query(Store).order_by(Store.id.asc()).all()
-        return render_template('items_update.html', item = item, stores = stores)
-    else: # POST
-        nom = request.form['nom']
-        unitats = int(request.form['unitats']) # es text, el passo a enter
-        store_id = int(request.form['store_id']) # es text, el passo a enter
 
-        # actualitzo els valors de l'item
-        item.nom = nom
-        item.unitats = unitats
-        item.store_id = store_id
+    # select que retorna una llista de resultats
+    stores = db.session.query(Store).order_by(Store.id.asc()).all()
+    
+    # creo el formulari amb les dades de l'item
+    form = ItemForm(obj = item)
+    form.store_id.choices = [(store.id, store.nom) for store in stores]
+    
+    if form.validate_on_submit(): # si s'ha fet submit al formulari
+        # dades del formulari a l'objecte item
+        form.populate_obj(item)
 
         # update!
         db.session.add(item)
@@ -42,30 +40,33 @@ def items_update(item_id):
 
         # https://en.wikipedia.org/wiki/Post/Redirect/Get
         return redirect(url_for('main_bp.items_read', item_id = item_id))
+    else: #GET
+        return render_template('items_update.html', item_id = item_id, form = form)
 
 @main_bp.route('/items/create', methods = ['POST', 'GET'])
 def items_create(): 
-    if request.method == 'GET':
-        # select que retorna una llista de resultats
-        stores = db.session.query(Store).order_by(Store.id.asc()).all()
-        return render_template('items_create.html', stores = stores)
-    else: # POST
-        nom = request.form['nom']
-        unitats = int(request.form['unitats']) # es text, el passo a enter
-        store_id = int(request.form['store_id']) # es text, el passo a enter
+    # select que retorna una llista de resultats
+    stores = db.session.query(Store).order_by(Store.id.asc()).all()
 
+    # creo el formulari buit
+    form = ItemForm()
+    form.store_id.choices = [(store.id, store.nom) for store in stores]
+
+    if form.validate_on_submit(): # si s'ha fet submit al formulari
         # he de crear un nou item
-        nou_item = Item()
-        nou_item.nom  = nom
-        nou_item.unitats  = unitats
-        nou_item.store_id = store_id
+        new_item = Item()
+        # dades del formulari a l'objecte item
+        form.populate_obj(new_item)
 
         # insert!
-        db.session.add(nou_item)
+        db.session.add(new_item)
         db.session.commit()
 
         # https://en.wikipedia.org/wiki/Post/Redirect/Get
         return redirect(url_for('main_bp.items_list'))
+    else: #GET
+        return render_template('items_create.html', form = form)
+
 
 @main_bp.route('/items/read/<int:item_id>')
 def items_read(item_id):
@@ -79,11 +80,12 @@ def items_delete(item_id):
     # select amb 1 resultat
     item = db.session.query(Item).filter(Item.id == item_id).one()
 
-    if request.method == 'GET':
-        return render_template('items_delete.html', item = item)
-    else: # POST
+    form = DeleteForm()
+    if form.validate_on_submit(): # si s'ha fet submit al formulari
         # delete!
         db.session.delete(item)
         db.session.commit()
 
         return redirect(url_for('main_bp.items_list'))
+    else: # GET
+        return render_template('items_delete.html', item = item, form = form)
